third_party/WebKit/LayoutTests/platform/linux/external/wpt/XMLHttpRequest/setrequestheader-case-insensitive-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: setRequestHeader() - headers that differ in case assert_equals: expected "x-test," but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"/b/c/b/linux_layout/src/third_party/WebKit/Tools/Scripts/webkitpy/thirdparty/wpt/wpt/tools/wptserve/wptserve/handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"/b/c/b/linux_layout/src/third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
Harness: the test ran to completion.
